refactor(models): move CacheToolsConfig to model.py
Address review feedback (#2232 review): define and export CacheToolsConfig
from model.py instead of bedrock.py so the dataclass lives next to
CacheConfig rather than in a model-provider-specific module.
